+++ src/notes_application_class.rb
