+++ test/test_note_application_class.rb
